14_Documentation/djlibrary/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-filter==2.4.0
django-taggit==0.22.2
djangorestframework==3.12.4
docutils==0.17.1
drf-yasg==1.20.0
idna==3.2
importlib-metadata==4.6.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
packaging==21.0
pkg-resources==0.0.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
zipp==3.5.0
